replace common memory classes
