--- src/ImageSharp/Memory/IManagedByteBuffer.cs
@@ -1,18 +1,0 @@
-// Copyright (c) Six Labors and contributors.
-// Licensed under the Apache License, Version 2.0.
-
-using System.Buffers;
-
-namespace SixLabors.Memory
-{
-    /// <summary>
-    /// Represents a byte buffer backed by a managed array. Useful for interop with classic .NET API-s.
-    /// </summary>
-    internal interface IManagedByteBuffer : IMemoryOwner<byte>
-    {
-        /// <summary>
-        /// Gets the managed array backing this buffer instance.
-        /// </summary>
-        byte[] Array { get; }
-    }
-}
