+++ tests/ImageSharp.Tests/Formats/Jpg/JpegColorConverterTests.cs
@@ -8,6 +8,7 @@ using SixLabors.ImageSharp.ColorSpaces;
 using SixLabors.ImageSharp.ColorSpaces.Conversion;
 using SixLabors.ImageSharp.Formats.Jpeg.Components.Decoder;
 using SixLabors.ImageSharp.Formats.Jpeg.Components.Decoder.ColorConverters;
+using SixLabors.ImageSharp.Tests.Memory;
 using SixLabors.Memory;
 
 using Xunit;
@@ -290,7 +291,8 @@ namespace SixLabors.ImageSharp.Tests.Formats.Jpg
                 }
 
                 // no need to dispose when buffer is not array owner
-                var source = new MemorySource<float>(new BasicArrayBuffer<float>(values), true);
+                var memory = new Memory<float>(values);
+                var source = new MemorySource<float>(memory);
                 buffers[i] = new Buffer2D<float>(source, values.Length, 1);
             }
             return new JpegColorConverter.ComponentValues(buffers, 0);
