+++ tests/ImageSharp.Tests/TestUtilities/TestMemoryAllocator.cs
@@ -18,14 +18,14 @@ namespace SixLabors.ImageSharp.Tests.Memory
         /// </summary>
         public byte DirtyValue { get; }
 
-        internal override IMemoryOwner<T> Allocate<T>(int length, AllocationOptions options = AllocationOptions.None)
+        public override IMemoryOwner<T> Allocate<T>(int length, AllocationOptions options = AllocationOptions.None)
         {
             T[] array = this.AllocateArray<T>(length, options);
 
             return new BasicArrayBuffer<T>(array, length);
         }
 
-        internal override IManagedByteBuffer AllocateManagedByteBuffer(int length, AllocationOptions options = AllocationOptions.None)
+        public override IManagedByteBuffer AllocateManagedByteBuffer(int length, AllocationOptions options = AllocationOptions.None)
         {
             byte[] array = this.AllocateArray<byte>(length, options);
             return new ManagedByteBuffer(array);
@@ -43,6 +43,70 @@ namespace SixLabors.ImageSharp.Tests.Memory
             }
 
             return array;
+        }
+
+        /// <summary>
+        /// Wraps an array as an <see cref="IManagedByteBuffer"/> instance.
+        /// </summary>
+        private class BasicArrayBuffer<T> : MemoryManager<T>
+            where T : struct
+        {
+            private GCHandle pinHandle;
+
+            /// <summary>
+            /// Initializes a new instance of the <see cref="BasicArrayBuffer{T}"/> class
+            /// </summary>
+            /// <param name="array">The array</param>
+            /// <param name="length">The length of the buffer</param>
+            public BasicArrayBuffer(T[] array, int length)
+            {
+                DebugGuard.MustBeLessThanOrEqualTo(length, array.Length, nameof(length));
+                this.Array = array;
+                this.Length = length;
+            }
+
+            /// <summary>
+            /// Initializes a new instance of the <see cref="BasicArrayBuffer{T}"/> class
+            /// </summary>
+            /// <param name="array">The array</param>
+            public BasicArrayBuffer(T[] array)
+                : this(array, array.Length)
+            {
+            }
+
+            /// <summary>
+            /// Gets the array
+            /// </summary>
+            public T[] Array { get; }
+
+            /// <summary>
+            /// Gets the length
+            /// </summary>
+            public int Length { get; }
+
+            /// <inheritdoc />
+            public override Span<T> GetSpan() => this.Array.AsSpan(0, this.Length);
+
+            public override unsafe MemoryHandle Pin(int elementIndex = 0)
+            {
+                if (!this.pinHandle.IsAllocated)
+                {
+                    this.pinHandle = GCHandle.Alloc(this.Array, GCHandleType.Pinned);
+                }
+
+                void* ptr = (void*)this.pinHandle.AddrOfPinnedObject();
+                return new MemoryHandle(ptr, this.pinHandle);
+            }
+
+            public override void Unpin()
+            {
+                throw new NotImplementedException();
+            }
+
+            /// <inheritdoc />
+            protected override void Dispose(bool disposing)
+            {
+            }
         }
 
         private class ManagedByteBuffer : BasicArrayBuffer<byte>, IManagedByteBuffer
